+++ src/commands/init/__test__/initCommand.test.ts
@@ -18,6 +18,7 @@ const copyFileSpy = jest.spyOn(fileHelper, 'copyFile');
 const createIndexSpy = jest.spyOn(fileHelper, 'createIndex');
 const createRouterSpy = jest.spyOn(fileHelper, 'createRouter');
 const createControllerSpy = jest.spyOn(fileHelper, 'createController');
+const createTestSpy = jest.spyOn(fileHelper, 'createTest');
 const adviceSpy = jest.spyOn(displayHelper, 'advice');
 
 const userConfig: UserAnswers = {
@@ -26,6 +27,7 @@ const userConfig: UserAnswers = {
   useTypescript: false,
   defaultPort: 3000,
   packageManager: 'npm',
+  useTest: false
 }
 
 describe('Test the initCommand function with no option', () => {
@@ -37,6 +39,7 @@ describe('Test the initCommand function with no option', () => {
     userConfig.useTypescript = false;
     userConfig.defaultPort = 3000;
     userConfig.packageManager = 'npm';
+    userConfig.useTest = false;
 
     jest.resetAllMocks();
   });
@@ -85,13 +88,17 @@ describe('Test the initCommand function with no option', () => {
 
     await initCommand(options);
 
-    expect(adviceSpy).toHaveBeenCalledTimes(2);
+    expect(adviceSpy).toHaveBeenCalledTimes(3);
     expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
       'npm install express dotenv cors',
     );
     expect(adviceSpy).toHaveBeenCalledWith(
       `Don't forget to install your dev dependencies with this command`,
       'npm install --save-dev nodemon @types/node @types/express ts-node typescript @types/cors',
+    );
+    expect(adviceSpy).toHaveBeenCalledWith(
+      'You can initialize the typescript config file with this command',
+      'npx tsc --init'
     );
   });
 
@@ -121,7 +128,7 @@ describe('Test the initCommand function with no option', () => {
 
     await initCommand(options);
 
-    expect(adviceSpy).toHaveBeenCalledTimes(2);
+    expect(adviceSpy).toHaveBeenCalledTimes(3);
     expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
       'npm install express dotenv ejs',
     );
@@ -154,21 +161,32 @@ describe('Test the initCommand function with no option', () => {
     userConfig.isApiRest = true;
     userConfig.hasViewEngine = true;
     userConfig.useTypescript = true;
+    userConfig.useTest = true;
 
     jest.spyOn(builder, 'getQuestion').mockResolvedValue(userConfig);
 
     await initCommand(options);
 
-    expect(createSpy).toHaveBeenCalledTimes(8);
+    expect(createSpy).toHaveBeenCalledTimes(9);
 
-    expect(adviceSpy).toHaveBeenCalledTimes(2);
+    expect(createTestSpy).toHaveBeenCalledTimes(1);
+
+    expect(adviceSpy).toHaveBeenCalledTimes(4);
     expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
       'npm install express dotenv ejs cors',
     );
     expect(adviceSpy).toHaveBeenCalledWith(
       `Don't forget to install your dev dependencies with this command`,
-      'npm install --save-dev nodemon @types/node @types/express ts-node typescript @types/cors',
+      'npm install --save-dev nodemon @types/node @types/express ts-node typescript @types/cors jest supertest @types/jest @types/supertest ts-jest',
     );
+    expect(adviceSpy).toHaveBeenCalledWith(
+      'You can initialize the typescript config file with this command',
+      'npx tsc --init'
+    );
+    expect(adviceSpy).toHaveBeenCalledWith(
+      `Don't forget to configure ts-jest for testing your app with jest and typescript. You can find more information here`,
+      'https://jestjs.io/docs/getting-started#using-typescript'
+    )
   });
 
   it('propose the yarn package manager', async () => {
@@ -178,13 +196,53 @@ describe('Test the initCommand function with no option', () => {
 
     await initCommand(options);
 
-    expect(adviceSpy).toHaveBeenCalledTimes(2);
+        expect(adviceSpy).toHaveBeenCalledTimes(2);
     expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
       'yarn add express dotenv',
     );
     expect(adviceSpy).toHaveBeenCalledWith(
       `Don't forget to install your dev dependencies with this command`,
       'yarn add --dev nodemon',
+    );
+  });
+
+  it('create an express application with test option available in javascript environment', async () => {
+    userConfig.useTest = true;
+
+    jest.spyOn(builder, 'getQuestion').mockResolvedValue(userConfig);
+
+    await initCommand(options);
+
+    expect(createSpy).toHaveBeenCalledTimes(6);
+
+    expect(createTestSpy).toHaveBeenCalledTimes(1);
+    expect(createTestSpy).toHaveBeenCalledWith('mainController.test', userConfig);
+
+    expect(adviceSpy).toHaveBeenCalledTimes(2);
+    expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
+      'npm install express dotenv',
+    );
+    expect(adviceSpy).toHaveBeenCalledWith(
+      `Don't forget to install your dev dependencies with this command`,
+      'npm install --save-dev nodemon jest supertest',
+    );
+  })
+
+  it('create an express application with test option available in typescript environment', async () => {
+    userConfig.useTest = true;
+    userConfig.useTypescript = true
+
+    jest.spyOn(builder, 'getQuestion').mockResolvedValue(userConfig);
+
+    await initCommand(options);
+
+    expect(adviceSpy).toHaveBeenCalledTimes(4);
+    expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
+      'npm install express dotenv',
+    );
+    expect(adviceSpy).toHaveBeenCalledWith(
+      `Don't forget to install your dev dependencies with this command`,
+      'npm install --save-dev nodemon @types/node @types/express ts-node typescript jest supertest @types/jest @types/supertest ts-jest',
     );
   })
 });
@@ -207,7 +265,7 @@ describe('Test the init command with the template option', () => {
       'npm install express dotenv cors',
     );
     expect(adviceSpy).toHaveBeenCalledWith(`Don't forget to install your dev dependencies with this command`,
-      'npm install --save-dev nodemon',
+      'npm install --save-dev nodemon jest supertest',
     );
   });
 
@@ -218,12 +276,12 @@ describe('Test the init command with the template option', () => {
 
     await initCommand(options);
 
-    expect(adviceSpy).toHaveBeenCalledTimes(2);
+    expect(adviceSpy).toHaveBeenCalledTimes(4);
     expect(adviceSpy).toHaveBeenCalledWith(`\nDon't forget to install your main dependencies with this command`,
       'npm install express dotenv cors',
     );
     expect(adviceSpy).toHaveBeenCalledWith(`Don't forget to install your dev dependencies with this command`,
-      'npm install --save-dev nodemon @types/node @types/express ts-node typescript @types/cors',
+      'npm install --save-dev nodemon @types/node @types/express ts-node typescript @types/cors jest supertest @types/jest @types/supertest ts-jest',
     );
   });
 })