+++ src/helpers/__test__/fileHelper.test.ts
@@ -17,6 +17,7 @@ const userConfiguration: UserConfiguration = {
   useTypescript: false,
   defaultPort: 3000,
   packageManager: 'npm',
+  useTest: false
 };
 
 const serverContent = `