+++ src/templates/controllers/controller.template.js
@@ -2,7 +2,7 @@
 
 const mainController = {
   home: function(request<% if (useTypescript) { %>: Request <% } %>, response<% if (useTypescript) { %>: Response <% } %>){
-    <% if(isApiRest) { %> response.json(`It's alive !`); <% } else { %>response.send(`<h1>It's alive !</h1>`); <% } %>
+    <% if(isApiRest) { %> response.json(`It's alive !`); <% } else { %>response.send(`It's alive !`); <% } %>
   }
 }